backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["stephen <chung25stephen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.59.9"
python-dotenv = "^1.0.1"
fastapi = {extras = ["standard"], version = "^0.115.6"}
requests = "^2.32.3"
bs4 = "^0.0.2"
pandas = "^2.2.3"
tiktoken = "^0.8.0"
transformers = "^4.48.1"
datasets = "^3.2.0"
torch = "^2.5.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.38"}
sqlmodel = "^0.0.22"
psycopg2 = "^2.9.10"
pymilvus = "^2.5.4"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
